requirements.txt
# ==== Core Frameworks ====
streamlit==1.43.2          # ✅ Khớp với version bạn đã cài qua pip show (bạn ghi 1.30.0 -> mình thay thành 1.43.2)
streamlit-option-menu==0.3.6
python-dotenv==0.21.1
supabase-py==1.0.3

# ==== AI / NLP / Deep Learning ====
openai>=1.3.0              # ⚠️ Giữ nguyên nhưng cần >=1.3.0 để dùng GPT-4, GPT-4o
tensorflow==2.15.0
keras==2.15.0              # ✅ Đồng bộ version với TensorFlow
torch==2.1.0
torchaudio==2.1.0

# ==== Machine Learning & Audio ====
librosa==0.10.1            # ⚠️ Yêu cầu numpy >=1.20.3 (bạn đang dùng 1.23.5 nên ✅)
pydub==0.25.1
soundfile==0.12.1
scikit-learn==1.3.0
nltk==3.8.1
numba==0.57
scipy==1.11.3

# ==== Visualization ====
matplotlib==3.8.2
seaborn==0.13.2

# ==== Image / Video Processing ====
opencv-python==4.7.0.72
rembg==2.0.50              # ✅ Dùng cho tách nền ảnh
Pillow==9.3.0

# ==== Misc & Utilities ====
bcrypt==4.0.1
numpy==1.23.5              # ⚠️ Nên đồng bộ theo yêu cầu của librosa, TensorFlow
pytube==15.0.0             # ✅ Dùng để download YouTube (cho AI Remix sau này)
requests==2.28.1           # ✅ Gọi các API như Suno AI, OpenAI...
